Fix social login in with automatic generate form html code
